refactor(mcp): move MCP data access into data/mcp; keep resource infra in lib
